moved rtsp teardown to client
